tests/wpt/tests/css/css-fonts/font-size-adjust-interpolation-math-functions.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>font-size-adjust interpolation with css math functions</title>
    <link rel="help" href="https://drafts.csswg.org/css-fonts-5/#font-size-adjust-prop">
    <link rel="help" href="https://drafts.csswg.org/css-values/#sign-funcs">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>

  <style>
    :root {
      font-size: 10px;
    }
  </style>
  <body>
    <script>
      test_interpolation({
        property: 'font-size-adjust',
        from: '100',
        to: 'calc(sign(20rem - 20px) * 180)',
      }, [
        {at: -1, expect: '20'},
        {at: 0, expect: '100'},
        {at: 0.125, expect: '110'},
        {at: 0.875, expect: '170'},
        {at: 1, expect: '180'},
        {at: 2, expect: '260'}
      ]);

      test_interpolation({
        property: 'font-size-adjust',
        from: 'calc(sign(20rem - 20px) * 100)',
        to: 'calc(sign(20rem - 20px) * 180)',
      }, [
        {at: -1, expect: '20'},
        {at: 0, expect: '100'},
        {at: 0.125, expect: '110'},
        {at: 0.875, expect: '170'},
        {at: 1, expect: '180'},
        {at: 2, expect: '260'}
      ]);
    </script>
  </body>
</html>
